requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
hikari==2.0.0.dev116
hikari-lightbulb==2.3.1
httplib2==0.21.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
pip==21.1.3
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
setuptools==56.0.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
